2025.1.70
2025.1.68
Bug fix

2025.1.66
Save power usage 24 and 48hours inbetween restarts, and reinstalls.

2025.1.65
Add device option to track on and off only if device currently using power.
Eg. switch is on - but not using power.
When starts using power will report this as on and time will be using power time.

2025.1.56
Add power usage